apps/api/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
langdetect>=1.0.9,<2.0.0
